+++ apiserver/facades/client/modelmanager/shims.go
@@ -6,6 +6,7 @@ package modelmanager
 import (
 	"github.com/juju/errors"
 	"github.com/juju/mgo/v2"
+	"github.com/juju/names/v4"
 	"github.com/juju/replicaset/v2"
 	"github.com/juju/version/v2"
 
@@ -14,6 +15,34 @@ import (
 	"github.com/juju/juju/environs/space"
 	"github.com/juju/juju/state"
 )
+
+// StatePool represents a point of use interface for getting the state from the
+// pool.
+type StatePool interface {
+	Get(string) (State, error)
+	MongoVersion() (string, error)
+}
+
+// State represents a point of use interface for modelling a current model.
+type State interface {
+	Model() (Model, error)
+	HasUpgradeSeriesLocks() (bool, error)
+	Release() bool
+	AllModelUUIDs() ([]string, error)
+	MachineCountForSeries(series ...string) (int, error)
+	MongoCurrentStatus() (*replicaset.Status, error)
+	SetModelAgentVersion(newVersion version.Number, stream *string, ignoreAgentVersions bool) error
+	AbortCurrentUpgrade() error
+}
+
+// Model defines a point of use interface for the model from state.
+type Model interface {
+	IsControllerModel() bool
+	AgentVersion() (version.Number, error)
+	Owner() names.UserTag
+	Name() string
+	MigrationMode() state.MigrationMode
+}
 
 type spaceStateShim struct {
 	common.ModelManagerBackend
